requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
astroid==4.0.2
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dill==0.4.0
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.4.11
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lomond==0.3.3
MarkupSafe==3.0.3
mccabe==0.7.0
narwhals==2.14.0
numpy==2.3.5
ollama==0.6.1
openai==2.17.0
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pylint==4.0.4
pypdf==6.4.2
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
